separate into analysis and run pass
